% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_name.R
\name{adjust_name}
\alias{adjust_name}
\title{Adjust a string by a symbol, motivated by fixing names of linkage groups for grepping (lg1 would grep lg10 as well)}
\usage{
adjust_name(name, length, symbol = "_")
}
\arguments{
\item{name}{a name to adjust}

\item{length}{a length of string to be extended by a symbol}

\item{symbol}{for extension. Default : '_'}
}
\description{
Adjust a string by a symbol, motivated by fixing names of linkage groups for grepping (lg1 would grep lg10 as well)
}
