% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asex_blue}
\alias{asex_blue}
\title{A blue colour used for asexual species}
\format{string}
\usage{
asex_blue
}
\description{
literally
asex_blue <- "#92C5DECD"
where first 6 hexadecimal digits after # are RGB intensities
and the last two hexadecimal digits is transparency (alpha)
}
\keyword{datasets}
