% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_mummer_coords.R
\name{load_mummer_coords}
\alias{load_mummer_coords}
\title{Load Mummer Coords file}
\usage{
load_mummer_coords(f, lg = NA, ord = NA)
}
\arguments{
\item{f}{a file that should be loaded}

\item{lg}{a linkage group that will be loaded (all of the linkage groups if lg = NA), specifying lg will speed up loading about ten fold. Default : NA}

\item{ord}{an order of scaffold that will be loaded, (all of the scaffolds groups if ord = NA). Specifying arguemtn has no effect on performance. Default : NA}
}
\description{
This function can filter theta dataframe by coverage (depth), window size or both
}
