% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_thetas.R
\name{load_thetas}
\alias{load_thetas}
\title{Load theta estimates}
\usage{
load_thetas(sp, ind, ref, window)
}
\arguments{
\item{sp}{- species}

\item{ind}{- individual}

\item{ref}{- refrence version ('b3v05')}

\item{window}{size of estimates}
}
\description{
This function allows you to load theta estimates from atlas
}
\examples{
Tdi_thetas <- load_thetas('1_Tdi', 'ref', 'b3v05', 100000)
}
