% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data_frame.R
\name{make_data_frame}
\alias{make_data_frame}
\title{Create a data.frame with defined variable names}
\usage{
make_data_frame(variables)
}
\arguments{
\item{variables}{a vector of veriable names}
}
\description{
Create a data.frame with defined variable names
}
