% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_log_hist.R
\name{plot_log_hist}
\alias{plot_log_hist}
\title{Plot log histogram with sex / asex overlaping data}
\usage{
plot_log_hist(sex_data, asex_data, xlim = NA, xlab = "Scafold size",
  breaks = 80, barwidth = 13, noborder = F, axes = T, cex.axis = 2,
  cex.lab = 1.6)
}
\arguments{
\item{sex_data}{a vector of sex data (must be numeric)}

\item{asex_data}{a vector of asex data}

\item{xlim}{[from data], xlab ['Scafold size'], cex.axis [2], cex.lab [1.6] inherited form \code{plot}}

\item{breaks}{[80] inherited from \code{hist}}

\item{barwidth}{width of histogram bars (I have not managed to make it perfectly automated) [13]}

\item{axes}{logical if to plot axis [T]}
}
\description{
Plot log histogram with sex / asex overlaping data
}
