% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_LaTeX_table.R
\name{print_LaTeX_table}
\alias{print_LaTeX_table}
\title{print LaTeX table}
\usage{
print_LaTeX_table(tabletoprint, alignment = "c")
}
\arguments{
\item{tabletoprint}{the table to be printed for LaTeX}

\item{alignment}{the aligment for the columns of the table (either vector of alignments or a single letter that will be used for all)}
}
\description{
This function prints a table that can be directly copy pasted into LaTeX
}
