% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sex_legend.R
\name{sex_legend}
\alias{sex_legend}
\title{Plot sex legend}
\usage{
sex_legend(pos = "topright", cex = 1.3)
}
\arguments{
\item{pos}{position of the legend ['topright']}

\item{cex}{the size of legend [1.3]}
}
\description{
This function plots a legend with sexual and sexual colours / symbols
}
