% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sex_red}
\alias{sex_red}
\title{A red colour used for sexual species}
\format{string}
\usage{
sex_red
}
\description{
literally
sex_red <- "#D6604DED"
where first 6 hexadecimal digits after # are RGB intensities
and the last two hexadecimal digits is transparency (alpha)
}
\keyword{datasets}
