% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_pairs}
\alias{species_pairs}
\title{Species pairs}
\format{a list of useful pair info
(pairs are sorted by age Timema pairs, mite pairs, Ostracod pairs)}
\usage{
species_pairs
}
\description{
DESC
}
\examples{
species_pairs$labels # returns a vector of pair labels
}
\keyword{datasets}
