% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssplit.R
\name{ssplit}
\alias{ssplit}
\title{Split a string by defined character and return unlisted vector of substrings}
\usage{
ssplit(s, split = "=")
}
\arguments{
\item{s}{a string to split}

\item{split}{a character used to split [default split = '=']}
}
\description{
Split a string by defined character and return unlisted vector of substrings
}
