% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timema_pairs}
\alias{timema_pairs}
\title{Timema species pairs}
\format{
a list
(pairs are sorted by divergence of Timema pairs)
}
\usage{
timema_pairs
}
\description{
a useful variables and information related to species pairs :
\itemize{
  \item `labels` - a vector of expressions;
  \item `JK-divergence` - divergences between pairs are calculated from transcriptomes;
}
}
\examples{
timema_pairs$labels # returns a vector of pair labels
}
\keyword{datasets}
