% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timemas}
\alias{timemas}
\title{Timema info}
\format{
a list
interleaved asexual and sexual sister species with pairs sorted by divergences
}
\usage{
timemas
}
\description{
a useful variables and information related to Timema species :
\itemize{
  \item `timema_codes`
  \item `timema_names`
  \item `labels` - a vector of expressions;
  \item `short_labels`
}
}
\examples{
timemas$labels # returns a vector of formated timema labels
}
\keyword{datasets}
