% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vioplot2.R
\name{vioplot2}
\alias{vioplot2}
\title{Plot smoothed distribution}
\usage{
vioplot2(
  x,
  ...,
  range = 1.5,
  h = NULL,
  ylim = NULL,
  names = NULL,
  horizontal = FALSE,
  col = "magenta",
  border = "black",
  lty = 1,
  lwd = 1,
  rectCol = "black",
  colMed = "white",
  pchMed = 19,
  at,
  add = FALSE,
  wex = 1,
  drawRect = TRUE,
  side = "both"
)
}
\arguments{
\item{x}{data vector to plot}

\item{...}{additional data vectors}

\item{range}{a coefficient to calculate the upper/lower adjacent values}

\item{h}{smoothing coefficient [estimated form data]}

\item{ylim, }{, rectCol, colMed, pchMed, drawRect - graphical details (mostly same as in default Graphics)}

\item{names}{one label, or a vector of labels for the datas must match the number of datas given}

\item{horizontal}{logical. horizontal or vertical distributions [F]}

\item{col, }{border, lty, lwd - Graphical parameters for the violin passed to lines and polygon}

\item{at}{where to plot distributions}

\item{add}{logical. if FALSE a new plot is created [F]}

\item{wex}{relative expansion of the distributions. [1]}

\item{side}{'left', 'right' or 'both' (violin plot). ['both']}
}
\description{
Plot smoothed distribution
}
