% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timemas}
\alias{timemas}
\title{Timema info}
\format{a list of useful Timema species info
interleaved asexual and sexual sister species}
\usage{
timemas
}
\description{
DESC
}
\examples{
timemas$labels # returns a vector of formated timema labels
}
\keyword{datasets}
